created export library test for method
removed all params as per new updated api route spec

added route in the rich-skill-service file
route will be /api/export/library. Created the component that will
call the route and return it and download it to users folder.

Button is now hidden when not authorized and visible when
user with proper role has access to it.
Added the path to routerLink so it just stays in the same page when clicked on

created button for library export
button checks auth roles to make sure only users with access can export