Upgrade base and build OSMT Docker images (#256)
* CentOS 8 is no longer supported and is at EOL
there is no CentOS 9 and they switched over to centOS stream which
i could not find in docker hub. I made a switch to rockylinux which is
one of the closes to centOS and created/maintained by one of the original creators
of centOS. Also i had to upgrade mvn to 3.8.6 because 3.8.4 dlcdn.apache.org no
longer had it available. gave me a 404 not found.

* switched to oraclelinux and tested it, works great

Co-authored-by: Edwin Santizo <edwin.santizo@wgu.edu>